+++ extensions/query-history/src/constants.ts
@@ -1,0 +1,10 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the Source EULA. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
+export const QUERY_HISTORY_CONFIG_SECTION = 'queryHistory';
+export const CAPTURE_ENABLED_CONFIG_SECTION = 'captureEnabled';
+export const DOUBLE_CLICK_ACTION_CONFIG_SECTION = 'doubleClickAction';
+
+export const ITEM_SELECTED_COMMAND_ID = 'queryHistory.itemSelected';