+++ extensions/sql-migration/src/models/stateMachine.ts
@@ -34,6 +34,11 @@ export enum State {
 	EXIT,
 }
 
+export enum ServiceTier {
+	GeneralPurpose = 'GeneralPurpose',
+	BusinessCritical = 'BusinessCritical',
+}
+
 export enum MigrationTargetType {
 	SQLVM = 'AzureSqlVirtualMachine',
 	SQLMI = 'AzureSqlManagedInstance',
@@ -194,6 +199,7 @@ export class MigrationStateModel implements Model, vscode.Disposable {
 	public _assessedDatabaseList!: string[];
 	public _runAssessments: boolean = true;
 	private _assessmentApiResponse!: mssql.AssessmentResult;
+	public _assessmentReportFilePath: string;
 	public mementoString: string;
 
 	public _databasesForMigration: string[] = [];
@@ -205,6 +211,7 @@ export class MigrationStateModel implements Model, vscode.Disposable {
 	public _skuRecommendationResults!: SkuRecommendation;
 	public _skuRecommendationPerformanceDataSource!: PerformanceDataSourceOptions;
 	private _skuRecommendationApiResponse!: mssql.SkuRecommendationResult;
+	public _skuRecommendationReportFilePaths: string[];
 	public _skuRecommendationPerformanceLocation!: string;
 	private _skuRecommendationRecommendedDatabaseList!: string[];
 	private _startPerfDataCollectionApiResponse!: mssql.StartPerfDataCollectionResult;
@@ -258,6 +265,8 @@ export class MigrationStateModel implements Model, vscode.Disposable {
 		this._databaseBackup.networkShares = [];
 		this._databaseBackup.blobs = [];
 		this._targetDatabaseNames = [];
+		this._assessmentReportFilePath = '';
+		this._skuRecommendationReportFilePaths = [];
 		this.mementoString = 'sqlMigration.assessmentResults';
 
 		this._skuScalingFactor = 100;
@@ -320,6 +329,7 @@ export class MigrationStateModel implements Model, vscode.Disposable {
 					}) ?? [],
 					errors: this._assessmentApiResponse?.errors ?? []
 				};
+				this._assessmentReportFilePath = response.assessmentReportPath;
 			} else {
 				this._assessmentResults = {
 					issues: [],
@@ -389,16 +399,19 @@ export class MigrationStateModel implements Model, vscode.Disposable {
 						sqlDbRecommendationResults: response?.sqlDbRecommendationResults ?? [],
 						sqlMiRecommendationResults: response?.sqlMiRecommendationResults ?? [],
 						sqlVmRecommendationResults: response?.sqlVmRecommendationResults ?? [],
-						instanceRequirements: response?.instanceRequirements
+						instanceRequirements: response?.instanceRequirements,
+						skuRecommendationReportPaths: response?.skuRecommendationReportPaths
 					},
 				};
+				this._skuRecommendationReportFilePaths = response.skuRecommendationReportPaths;
 			} else {
 				this._skuRecommendationResults = {
 					recommendations: {
 						sqlDbRecommendationResults: [],
 						sqlMiRecommendationResults: [],
 						sqlVmRecommendationResults: [],
-						instanceRequirements: response?.instanceRequirements
+						instanceRequirements: response?.instanceRequirements,
+						skuRecommendationReportPaths: response?.skuRecommendationReportPaths
 					},
 				};
 			}
@@ -411,7 +424,8 @@ export class MigrationStateModel implements Model, vscode.Disposable {
 					sqlDbRecommendationResults: this._skuRecommendationApiResponse?.sqlDbRecommendationResults ?? [],
 					sqlMiRecommendationResults: this._skuRecommendationApiResponse?.sqlMiRecommendationResults ?? [],
 					sqlVmRecommendationResults: this._skuRecommendationApiResponse?.sqlVmRecommendationResults ?? [],
-					instanceRequirements: this._skuRecommendationApiResponse?.instanceRequirements
+					instanceRequirements: this._skuRecommendationApiResponse?.instanceRequirements,
+					skuRecommendationReportPaths: this._skuRecommendationApiResponse?.skuRecommendationReportPaths
 				},
 				recommendationError: error
 			};