+++ i18n/ads-language-pack-it/CHANGELOG.md
@@ -2,6 +2,8 @@
 Tutti i cambiamenti degni di nota al language pack "ads-language-pack-it" saranno documentati in questo file.
 
 ## [Released]
+* August 12, 2022 - Release for Azure Data Studio 1.39
+* July 20, 2022 - Release for Azure Data Studio 1.38
 * June 3, 2022 - Release for Azure Data Studio 1.37
 * April 6, 2022 - Release for Azure Data Studio 1.36
 * February 23, 2022 - Release for Azure Data Studio 1.35